refactor cancerspecif
