+++ immunopepper/mode_cancerspecif.py
@@ -1,25 +1,23 @@
 import logging
 import os
-import sys
 
-
-from immunopepper.spark_config import create_spark_session_from_config
-from immunopepper.spark_config import default_spark_config
-from immunopepper.spark import check_interm_files
+from immunopepper.sdisk import check_interm_files
+from immunopepper.sdisk import output_count
+from immunopepper.sdisk import redirect_interm
+from immunopepper.sdisk import save_output_count
+from immunopepper.sdisk import save_spark
+from immunopepper.sloaders import process_build_outputs
+from immunopepper.sloaders import process_libsize
+from immunopepper.sloaders import remove_external_kmer_list
 from immunopepper.spark import combine_cancer
 from immunopepper.spark import combine_hard_threshold_cancers
 from immunopepper.spark import combine_hard_threshold_normals
 from immunopepper.spark import combine_normals
 from immunopepper.spark import filter_expr_kmer
 from immunopepper.spark import filter_hard_threshold
-from immunopepper.spark import output_count
-from immunopepper.spark import process_build_outputs
-from immunopepper.spark import process_libsize
-from immunopepper.spark import redirect_interm
-from immunopepper.spark import remove_external_kmer_list
 from immunopepper.spark import remove_uniprot
-from immunopepper.spark import save_output_count
-from immunopepper.spark import save_spark
+from immunopepper.spark_config import create_spark_session_from_config
+from immunopepper.spark_config import default_spark_config
 from pyspark.sql.functions import broadcast
